Add further tests around deleting a user; You cannot delete a user that does not exist; You must be logged in to delete a user; You can only delete your own account

+++ apps/backend/src/user/user-router.ts
@@ -90,9 +90,14 @@ const loginRequestHandlerFactory =
 const deleteRequestHandlerFactory =
   (userService: UserService): RequestHandler =>
   async (req, res, next) => {
+    const authenticatedUsersEmail = res.locals.claims.email;
     const { email } = req.body;
-    const response = await userService.delete(email);
-    res.status(200).send(response);
+    if (authenticatedUsersEmail === email) {
+      const response = await userService.delete(email);
+      res.status(200).send(response);
+    } else {
+      res.status(200).send({ isSuccess: false });
+    }
   };
 
 const userRouterFactory = (
